+++ man-roxygen/commonargs.R
@@ -1,1 +1,6 @@
 #' @param popid (character or number) A NMFS_POPID for a population
+#' @param extra extra query params or column values. Enter as a list, e.g. list(pageno=7)
+#' @param table_id not normally needed as this is set automatically to the table id for NOSA using the table saved from `rcax_table()`. But you can set it if needed.
+#' @param cols column names to return. Use cols=NULL if you want all columns
+
+